git-svn-id: svn://localhost/mosa@821 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/HelloWorld/Memory.cs
@@ -78,7 +78,7 @@ namespace Mosa.Kernel.Memory.X86
 		/// </summary>
 		/// <param name="location">The location.</param>
 		/// <returns></returns>
-		public unsafe static ulong Get64(uint location)
+		public unsafe static ulong Get64(ulong location)
 		{
 			ulong* at = (ulong*)location;
 			return *at;
